Update patch set 1
Patch Set 1:

Have you discussed with the heat-translator team about these NFV features? Perhaps it's best to agree in advance that some code will be out of the master branch. In such case, have you also considered including heat-translator as a git submodule? This will let you be in sync with the project's origin instead of forking it.

Patch-set: 1