Provide a specific stack_name when deploy a template use heat-translator
Currently, the stack ame of a template is random when use
heat-translator online, and there is not any other feedback information
except success or failure, the result is not very convenient.
So it is necessary to provide a stack name when deploy a template.

Note: the improvement will submit to openstack community.

JIRA:PARSER-91

Change-Id: Ia5f99c38301211dfb9c21a8b10dd31a50b82f966
Signed-off-by: shangxdy <shang.xiaodong@zte.com.cn>