Update patch set 1
Patch Set 1:

HI, howard, the scripts are simple currently only for validation the framework using nfv tosca, and the contents will not be about vendor specific. And the plan is to validate network(such as sriov) or vm(such as affinity and anti-affinity), which are add in the future.

Patch-set: 1
Label: Verified=0