Fix BlockStorageAttachment bug when translate compute node
Currently, when translate compute node with local_storage and
high_availability, there is not BlockStorageAttache resource created
about block storage, and only BlockStorage created.
If only local_storage requirement in compute node, the result is
correct.
The reason is the function logic of _get_attachment_node is error. and
the patch will fixit.

JIRA:PARSER-100

Change-Id: I36456089ae65102d4b17797d266c850d8c16b29d
Signed-off-by: shangxdy <shang.xiaodong@zte.com.cn>